build: Fix circular dependency
